error messages updated

+++ HNTAS/HNTAS.Web.UI/Models/NetworkElements/NetworkElementViewModel.cs
@@ -9,7 +9,7 @@ namespace HNTAS.Web.UI.Models.NetworkElements
 
         [MustHaveOneItem(ErrorMessage = "Select at least one element that is part of your heat network")]
         public List<HeatNetworkElementType> SelectedElementIds { get; set; } = new();
-        public Dictionary<HeatNetworkElementType, int?> ElementCounts { get; set; } = new();
+        public Dictionary<HeatNetworkElementType, double?> ElementCounts { get; set; } = new();
     }
 
     public class NetworkElementOption
